feat: check password before editing note

+++ scripts/new.sh
@@ -7,6 +7,12 @@ PASSWORD=$(echo $PASSWORD | base64)
 
 echo ""
 
+RESP=$(curl -s -X POST -H "Authorization: Basic $PASSWORD" $HOST/api/check-pwd)
+if [ "$RESP" != "ok" ]; then
+  echo "invalid password"
+  exit 1
+fi
+
 rm -f .nts-note
 touch .nts-note
 ${EDITOR:-nano} .nts-note
